LDE-714: added further read tests
