+++ kartothek/io/testing/read.py
@@ -29,7 +29,6 @@ The following fixtures should be present (see tests.read.conftest)
 
 import datetime
 from functools import partial
-from itertools import permutations
 
 import pandas as pd
 import pandas.testing as pdt
@@ -132,217 +131,9 @@ def test_store_input_types(store_input_types, bound_load_dataframes):
     pdt.assert_frame_equal(result, df.head(1), check_dtype=False)
 
 
-def _perform_read_test(
-    dataset_uuid,
-    store_factory,
-    execute_read_callable,
-    use_categoricals,
-    output_type,
-    dates_as_object,
-    read_kwargs=None,
-    ds_factory=None,
-):
-    if not read_kwargs:
-        read_kwargs = {}
-    if use_categoricals:
-        # dataset_with_index has an index on L but not on P
-        categoricals = ["P", "L"]
-    else:
-        categoricals = None
-
-    result = execute_read_callable(
-        dataset_uuid=dataset_uuid,
-        store=store_factory,
-        factory=ds_factory,
-        categoricals=categoricals,
-        dates_as_object=dates_as_object,
-        **read_kwargs,
-    )
-
-    assert len(result) == 2
-
-    if output_type == "metapartition":
-        for res in result:
-            assert isinstance(res, MetaPartition)
-        result = [mp.data for mp in result]
-
-        def sort_by(obj):
-            return obj[SINGLE_TABLE].P.iloc[0]
-
-    elif output_type == "table":
-        assert isinstance(result[0], pd.DataFrame)
-        assert "P" in result[0]
-
-        def sort_by(obj):
-            return obj.P.iloc[0]
-
-    else:
-        assert isinstance(result[0], pd.DataFrame)
-        assert "P" in result[0]
-
-        def sort_by(obj):
-            return obj.P.iloc[0]
-
-    result = sorted(result, key=sort_by)
-
-    expected_df_core_1 = pd.DataFrame(
-        {"P": [1], "L": [1], "TARGET": [1], "DATE": [datetime.date(2010, 1, 1)]}
-    )
-    expected_df_core_2 = pd.DataFrame(
-        {"P": [2], "L": [2], "TARGET": [2], "DATE": [datetime.date(2009, 12, 31)]}
-    )
-    expected_dfs = [
-        expected_df_core_1,
-        expected_df_core_2,
-    ]
-
-    for res, expected_df_core in zip(result, expected_dfs):
-        if not dates_as_object:
-            expected_df_core["DATE"] = pd.to_datetime(expected_df_core["DATE"])
-        if use_categoricals:
-            expected_df_core = expected_df_core.astype(
-                {"P": "category", "L": "category"}
-            )
-
-        pdt.assert_frame_equal(
-            res.reset_index(drop=True),
-            expected_df_core.reset_index(drop=True),
-            check_dtype=False,
-            check_like=True,
-            check_categorical=False,
-        )
-
-
-@pytest.mark.parametrize(
-    "predicates",
-    [
-        [[("P", "==", 2)]],
-        [[("P", "in", [2])]],
-        [[("P", "!=", 1)]],
-        [[("P", ">", 1)]],
-        [[("P", ">=", 2)]],
-    ],
-)
-def test_read_dataset_as_dataframes_predicate(
-    dataset, store_session_factory, bound_load_dataframes, predicates, output_type
-):
-    if output_type != "dataframe":
-        pytest.skip()
-    result = bound_load_dataframes(
-        dataset_uuid=dataset.uuid, store=store_session_factory, predicates=predicates,
-    )
-    core_result = pd.concat(result)
-
-    expected_core = pd.DataFrame(
-        {"P": [2], "L": [2], "TARGET": [2], "DATE": [datetime.date(2009, 12, 31)]}
-    )
-    pdt.assert_frame_equal(
-        core_result, expected_core, check_dtype=False, check_like=True
-    )
-
-
-@pytest.mark.parametrize(
-    "predicates",
-    [
-        [[("P", "==", 2), ("TARGET", "==", 2)]],
-        [[("P", "in", [2]), ("TARGET", "==", 2)]],
-        [[("P", "!=", 1), ("L", "==", 2)]],
-        [[("P", "!=", 1), ("L", "in", [2])]],
-        [[("P", ">", 2)], [("TARGET", ">=", 2)]],
-        [[("P", ">=", 2)], [("TARGET", ">=", 2)]],
-    ],
-)
-def test_read_dataset_as_dataframes_predicate_with_partition_keys(
-    dataset_partition_keys,
-    store_session_factory,
-    bound_load_dataframes,
-    predicates,
-    output_type,
-):
-    if output_type != "dataframe":
-        pytest.skip()
-    result = bound_load_dataframes(
-        dataset_uuid=dataset_partition_keys.uuid,
-        store=store_session_factory,
-        predicates=predicates,
-    )
-
-    core_result = pd.concat(result)
-
-    expected_core = pd.DataFrame(
-        {"P": [2], "L": [2], "TARGET": [2], "DATE": [datetime.date(2009, 12, 31)]}
-    )
-    pdt.assert_frame_equal(
-        core_result, expected_core, check_dtype=False, check_like=True
-    )
-
-
-def test_read_dataset_as_dataframes_predicate_empty(
-    dataset_partition_keys, store_session_factory, output_type, bound_load_dataframes,
-):
-    if output_type != "dataframe":
-        pytest.skip()
-    result = bound_load_dataframes(
-        dataset_uuid=dataset_partition_keys.uuid,
-        store=store_session_factory,
-        predicates=[[("P", "==", -42)]],
-        columns={SINGLE_TABLE: ["P", "L", "TARGET"]},
-    )
-    assert len(result) == 0
-
-
 def _gen_partition(b_c):
     b, c = b_c
     return pd.DataFrame({"a": [1], "b": [b], "c": c})
-
-
-def test_read_dataset_as_dataframes_dispatch_by_empty(
-    store_session_factory,
-    dataset_dispatch_by,
-    bound_load_dataframes,
-    backend_identifier,
-    output_type,
-    metadata_version,
-    dataset_dispatch_by_uuid,
-):
-    if output_type == "table":
-        pytest.skip()
-    # Dispatch by primary index "A"
-    dispatched = bound_load_dataframes(
-        dataset_uuid=dataset_dispatch_by_uuid,
-        store=store_session_factory,
-        dispatch_by=[],
-    )
-
-    assert len(dispatched) == 1
-
-
-@pytest.mark.parametrize("dispatch_by", ["A", "B", "C"])
-def test_read_dataset_as_dataframes_dispatch_by_single_col(
-    store_session_factory,
-    dataset_dispatch_by,
-    bound_load_dataframes,
-    backend_identifier,
-    dispatch_by,
-    output_type,
-    metadata_version,
-    dataset_dispatch_by_uuid,
-):
-    if output_type == "table":
-        pytest.skip()
-    # Dispatch by primary index "A"
-    dispatched_a = bound_load_dataframes(
-        dataset_uuid=dataset_dispatch_by_uuid,
-        store=store_session_factory,
-        dispatch_by=[dispatch_by],
-    )
-
-    unique_a = set()
-    for data in dispatched_a:
-        unique_dispatch = data[dispatch_by].unique()
-        assert len(unique_dispatch) == 1
-        assert unique_dispatch[0] not in unique_a
-        unique_a.add(unique_dispatch[0])
 
 
 @pytest.fixture(scope="session")
@@ -377,101 +168,6 @@ def dataset_dispatch_by(
         secondary_indices=["C"],
     )
     return pd.concat(clusters).sort_values(["A", "B", "C"]).reset_index(drop=True)
-
-
-def test_read_dataset_as_dataframes_dispatch_by_multi_col(
-    store_session_factory,
-    bound_load_dataframes,
-    output_type,
-    dataset_dispatch_by,
-    dataset_dispatch_by_uuid,
-):
-    if output_type == "table":
-        pytest.skip()
-    for dispatch_by in permutations(("A", "B", "C"), 2):
-        dispatched = bound_load_dataframes(
-            dataset_uuid=dataset_dispatch_by_uuid,
-            store=store_session_factory,
-            dispatch_by=dispatch_by,
-        )
-        uniques = pd.DataFrame(columns=dispatch_by)
-        for part in dispatched:
-            if isinstance(part, MetaPartition):
-                data = part.data
-            else:
-                data = part
-            unique_dispatch = data[list(dispatch_by)].drop_duplicates()
-            assert len(unique_dispatch) == 1
-            row = unique_dispatch
-            uniques.append(row)
-        assert not any(uniques.duplicated())
-
-
-@pytest.mark.parametrize(
-    "dispatch_by, predicates, expected_dispatches",
-    [
-        # This should only dispatch one partition since there is only
-        # one file with valid data points
-        (["A"], [[("C", ">", 2)]], 1),
-        # We dispatch and restrict to one valie, i.e. one dispatch
-        (["B"], [[("B", "==", 10)]], 1),
-        # The same is true for a non-partition index col
-        (["C"], [[("C", "==", 1)]], 1),
-        # A condition where both primary and secondary indices need to work together
-        (["A", "C"], [[("A", ">", 1), ("C", "<", 3)]], 2),
-    ],
-)
-def test_read_dispatch_by_with_predicates(
-    store_session_factory,
-    dataset_dispatch_by_uuid,
-    bound_load_dataframes,
-    dataset_dispatch_by,
-    dispatch_by,
-    output_type,
-    expected_dispatches,
-    predicates,
-):
-    if output_type == "table":
-        pytest.skip()
-
-    dispatched = bound_load_dataframes(
-        dataset_uuid=dataset_dispatch_by_uuid,
-        store=store_session_factory,
-        dispatch_by=dispatch_by,
-        predicates=predicates,
-    )
-
-    assert len(dispatched) == expected_dispatches, dispatched
-
-
-def test_read_dataset_as_dataframes(
-    dataset,
-    store_session_factory,
-    dataset_factory,
-    use_dataset_factory,
-    bound_load_dataframes,
-    use_categoricals,
-    output_type,
-    dates_as_object,
-):
-    if use_dataset_factory:
-        dataset_uuid = dataset.uuid
-        store_factory = store_session_factory
-        ds_factory = None
-    else:
-        dataset_uuid = None
-        store_factory = None
-        ds_factory = dataset_factory
-
-    _perform_read_test(
-        dataset_uuid=dataset_uuid,
-        store_factory=store_factory,
-        ds_factory=ds_factory,
-        execute_read_callable=bound_load_dataframes,
-        use_categoricals=use_categoricals,
-        output_type=output_type,
-        dates_as_object=dates_as_object,
-    )
 
 
 def test_read_dataset_as_dataframes_columns_projection(
